Refactors and moves around utility functions in cli
